Renamed project to align with COVID Green and updated android/ios builds to work with new name
