+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.covidtracker.hse",
+    package = "io.lfph.covidgreen",
 )
 
 android_resource(
     name = "res",
-    package = "com.covidtracker.hse",
+    package = "io.lfph.covidgreen",
     res = "src/main/res",
 )
