+++ android/build.gradle
@@ -13,8 +13,7 @@ buildscript {
         maven { url 'https://dl.bintray.com/android/android-tools/' }
     }
     dependencies {
-        classpath('com.android.tools.build:gradle:3.6.3')
-
+        classpath('com.android.tools.build:gradle:4.0.0')
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
