+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'COVIDTracker'
+rootProject.name = 'COVIDGreen'
 apply from: '../node_modules/react-native-unimodules/gradle.groovy'
 includeUnimodulesProjects()
 apply from: file("../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesSettingsGradle(settings)
