+++ assets/lang/en.json
@@ -1,8 +1,8 @@
 {
   "common": {
-    "name": "COVID Tracker Ireland",
+    "name": "COVID Green",
     "url": "https://www.covidtracker.ie/",
-    "message": "Join COVID Tracker Ireland to fight against COVID-19",
+    "message": "Join COVID Green to fight against COVID-19",
     "networkError": "Connection issue. Please check your internet connection and try again.",
     "missingDataAction": "Refresh page",
     "missingError": "Loading issue. We were not able to load some of the data correctly.",
@@ -25,7 +25,7 @@
     "dataPolicy": "Data Protection Policy",
     "terms": "Terms and Conditions",
     "metrics":"Settings For App Metrics",
-    "leave": "Leave COVID Tracker Ireland",
+    "leave": "Leave COVID Green",
     "contactTracing": "Contact Tracing"
   },
   "navbar": {
@@ -43,7 +43,7 @@
   },
   "onboarding": {
     "intro": {
-      "title": "With COVID Tracker you can do 3 things to help fight COVID-19:",
+      "title": "With COVID Green you can do 3 things to help fight COVID-19:",
       "view1": {
         "title": "Contact Tracing",
         "text": "Help contact tracing, protecting your privacy."
@@ -181,7 +181,7 @@
   },
   "leave": {
     "title": "Leave",
-    "info": "Thank you for helping us fight COVID-19. We're sorry to see you go. When you tap 'I want to leave' we will remove all data stored by the app from your device including a mobile number, if you have shared one, symptom data and any demographic health data you may have shared. Non-identifying authorisation tokens stored on the server will also be deleted.\n\nRandom IDs created or collected by Exposure Notification Services cannot be removed by the COVID Tracker app. If you wish to remove these Random IDs you can do this via your device Settings.",
+    "info": "Thank you for helping us fight COVID-19. We're sorry to see you go. When you tap 'I want to leave' we will remove all data stored by the app from your device including a mobile number, if you have shared one, symptom data and any demographic health data you may have shared. Non-identifying authorisation tokens stored on the server will also be deleted.\n\nRandom IDs created or collected by Exposure Notification Services cannot be removed by the COVID Green app. If you wish to remove these Random IDs you can do this via your device Settings.",
     "button": "I want to leave",
     "confirmTitle": "Leave",
     "confirmText": "Are you sure you want to leave?",
@@ -302,7 +302,7 @@
   },
   "yourData": {
     "title": "Your data",
-    "info": "The COVID Tracker app protects your privacy and does not share personal information about you with other app users. **Your identity will never be revealed to other app users.**\n\nAny personal data you provide will be processed in line with GDPR and data protection law. You can read more about this in the Data Protection Information Notice below. **Your data will only be used in relation to the COVID-19 pandemic response** as set out in the Data Protection Information Notice.\n\nIf you need to be alerted, the app will start this process with a secure in-app notification. Separate to using this app, the HSE contact tracing team may phone you if someone with COVID-19 identifies you as a close contact. If you have a COVID-19 test, the HSE will contact you by text or by phone with results.\n\nTake care with any suspicious phone calls, emails or texts asking you for personal information. **The HSE will not ask you for personal information by text or email.**",
+    "info": "The COVID Green app protects your privacy and does not share personal information about you with other app users. **Your identity will never be revealed to other app users.**\n\nAny personal data you provide will be processed in line with GDPR and data protection law. You can read more about this in the Data Protection Information Notice below. **Your data will only be used in relation to the COVID-19 pandemic response** as set out in the Data Protection Information Notice.\n\nIf you need to be alerted, the app will start this process with a secure in-app notification. Separate to using this app, the HSE contact tracing team may phone you if someone with COVID-19 identifies you as a close contact. If you have a COVID-19 test, the HSE will contact you by text or by phone with results.\n\nTake care with any suspicious phone calls, emails or texts asking you for personal information. **The HSE will not ask you for personal information by text or email.**",
     "viewInSettings": "You can view this notice at any time in Settings. If you do not want to continue, close and delete the app. If you consent, tap Continue to proceed.",
     "continue": "Continue"
   },
@@ -332,7 +332,7 @@
   },
   "closeContactNotification": {
     "title": "Close Contact Alert",
-    "description": "Please check the COVID Tracker app now."
+    "description": "Please check the COVID Green app now."
   },
   "closeContactWarn": {
     "title": "Close Contact Alert",
@@ -342,7 +342,7 @@
     "title": "Close Contact Alert",
     "infoTitle": "Close contact information",
     "intro": "If you have received a notification from the app, or a call from our Contact Tracing team, telling you that you have been in close contact with someone that has tested positive for coronavirus, you may need to be tested.",
-    "alertintro": "The COVID Tracker App has detected that you have been in close contact with someone that has tested positive for coronavirus. You may need to be tested.",
+    "alertintro": "The COVID Green App has detected that you have been in close contact with someone that has tested positive for coronavirus. You may need to be tested.",
     "callBack": "Call Back",
     "callBackQueue": "The HSE contact tracing team will call you as soon as possible.",
     "todo": {
