+++ components/views/contact-tracing/not-active.tsx
@@ -46,7 +46,7 @@ export const NotActive: FC<NotActiveProps> = ({
         } else {
           await IntentLauncher.startActivityAsync(
             IntentLauncher.ACTION_APPLICATION_DETAILS_SETTINGS,
-            {data: 'package:com.covidtracker.hse'}
+            {data: 'package:io.lfph.covidgreen'}
           );
         }
