+++ components/views/settings/contact-tracing.tsx
@@ -67,7 +67,7 @@ export const ContactTracingSettings = () => {
         } else {
           await IntentLauncher.startActivityAsync(
             IntentLauncher.ACTION_APPLICATION_DETAILS_SETTINGS,
-            {data: 'package:com.covidtracker.hse'}
+            {data: 'package:io.lfph.covidgreen'}
           );
         }
         await supportsExposureApi();
