+++ docs/technical/Readme.md
@@ -1,19 +1,19 @@
-# COVID Tracker Architecture
-The following section describes the COVID Tracker App architecture. 
+# COVID Green Architecture
+The following section describes the COVID Green App architecture. 
 
 ## App Flows
-- [COVID Tracker Onboarding Flow](https://github.com/HSEIreland/covidtracker-documentation/blob/master/documentation/technical/covid-tracker-onboarding-flow.md) - describes the onboarding process for new users 
-- [COVID Tracker News & Information Flow](https://github.com/HSEIreland/covidtracker-documentation/blob/master/documentation/technical/covid-tracker-newsinfo-flow.md) - describes the News & Information aspects of the app
-- [COVID Tracker Day to Day Contact Tracing Flow](https://github.com/HSEIreland/covidtracker-documentation/blob/master/documentation/technical/covid-tracker-d2d-ct-exposure-event.md) - describes day to day contact tracing
-- [COVID Tracker Positive Test Result Flow](https://github.com/HSEIreland/covidtracker-documentation/blob/master/documentation/technical/covid-tracker-positive-test-result-flow.md) - describes a the flow for a positive test result
-- [COVID Tracker Check-in Flow](https://github.com/HSEIreland/covidtracker-documentation/blob/master/documentation/technical/covid-tracker-check-in-onboarding-flow.md) - describes the COVID Check-In 
-- [COVID Tracker Leave Flow](https://github.com/HSEIreland/covidtracker-documentation/blob/master/documentation/technical/covid-tracker-leave-flow.md)  - describes the Leave flow
+- [COVID Green Onboarding Flow](https://github.com/covidgreen/covid-green-app/blob/current/docs/technical/covid-green-onboarding-flow.md) - describes the onboarding process for new users 
+- [COVID Green News & Information Flow](https://github.com/covidgreen/covid-green-app/blob/current/docs/technical/covid-green-newsinfo-flow.md) - describes the News & Information aspects of the app
+- [COVID Green Day to Day Contact Tracing Flow](https://github.com/covidgreen/covid-green-app/blob/current/docs/technical/covid-green-d2d-ct-exposure-event.md) - describes day to day contact tracing
+- [COVID Green Positive Test Result Flow](https://github.com/covidgreen/covid-green-app/blob/current/docs/technical/covid-green-positive-test-result-flow.md) - describes a the flow for a positive test result
+- [COVID Green Check-in Flow](https://github.com/covidgreen/covid-green-app/blob/current/docs/technical/covid-green-check-in-onboarding-flow.md) - describes the COVID Check-In 
+- [COVID Green Leave Flow](https://github.com/covidgreen/covid-green-app/blob/current/docs/technical/covid-green-leave-flow.md)  - describes the Leave flow
 
 ## App Backend Services
-- [COVID Tracker Backend Services Views](https://github.com/HSEIreland/covidtracker-documentation/blob/master/documentation/technical/covid-tracker-backend-views.md) - describes the backed services, APIs, Services, AWS environment
-- [COVID Tracker Authentication Scheme](https://github.com/HSEIreland/covidtracker-documentation/blob/master/documentation/technical/covid-tracker-auth.md) - describes the app authentication scheme
+- [COVID Green Backend Services Views](https://github.com/covidgreen/covid-green-app/blob/current/docs/technical/covid-green-backend-views.md) - describes the backed services, APIs, Services, AWS environment
+- [COVID Green Authentication Scheme](https://github.com/covidgreen/covid-green-app/blob/current/docs/technical/covid-green-auth.md) - describes the app authentication scheme
 
 ## Data
-- [COVID Tracker App Data View](https://github.com/HSEIreland/covidtracker-documentation/blob/master/documentation/technical/covid-tracker-app-data.md) - describes the data in the app on the users device
-- [COVID Tracker App Backend Data Views](https://github.com/HSEIreland/covidtracker-documentation/blob/master/documentation/technical/covid-tracker-backend-data.md) - describes the data used in the backend 
-- [COVID Tracker App Metrics View](https://github.com/HSEIreland/covidtracker-documentation/blob/master/documentation/technical/covid-tracker-app-metrics.md) - describes the metrics collected by the app
+- [COVID Green App Data View](https://github.com/covidgreen/covid-green-app/blob/current/docs/technical/covid-green-app-data.md) - describes the data in the app on the users device
+- [COVID Green App Backend Data Views](https://github.com/covidgreen/covid-green-app/blob/current/docs/technical/covid-green-backend-data.md) - describes the data used in the backend 
+- [COVID Green App Metrics View](https://github.com/covidgreen/covid-green-app/blob/current/docs/technical/covid-green-app-metrics.md) - describes the metrics collected by the app
