+++ docs/technical/covid-green-app-metrics.md
@@ -1,4 +1,4 @@
-# Core COVID Tracker App indicator set
+# Core COVID Green App indicator set
 The indicators cover 4 areas: adoption and use; contact tracing; symptom tracking; and, epidemiology. The purpose of the first three areas is performance monitoring, while the fourth is analysis and modelling. Data is measured daily for each indicator and can be reported daily, cumulatively, or as events occur.
 | **Indicator area**         | **Indicator focus**                  | **Output indicator**                                                                    | **Description**                                                                                                                                                                                                                                                                   |
 |----------------------------|--------------------------------------|-----------------------------------------------------------------------------------------|-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
