+++ docs/technical/covid-green-auth.md
@@ -1,2 +1,2 @@
-# COVID Tracker Authentication Scheme
+# COVID Green Authentication Scheme
 ![Authentication Scheme](https://app.lucidchart.com/publicSegments/view/f5d23a46-a3fa-48cb-94c6-9861e9d794f7/image.png)
