+++ docs/technical/covid-green-backend-views.pdf
