+++ ios/COVIDGreen/AppDelegate.h
