+++ ios/COVIDGreen/Base.lproj/LaunchScreen.xib
