+++ ios/COVIDGreen/COVIDGreen.entitlements
