+++ ios/COVIDGreen/en.lproj/LaunchScreen.strings
