+++ ios/COVIDGreen/ga.lproj/LaunchScreen.strings
